feat: implement smart 13th week detection and refactor cache operations
  - Switch from /data/seasons/list to /data/series/get for better series metadata
  - Add intelligent 13th week series detection and conditional data fetching
  - Enhance database schema with new fields: eligible, categoryId, isSpecialEvent,
  specialEventType, maxStarters, minStarters
  - Implement graceful error handling for "No chunk file names found" API responses
  - Refactor cache operations into modular structure:
    - cache/series.ts - Series caching logic
    - cache/weekly-results.ts - Weekly results with 13th week filtering
    - cache/seasons-schedule.ts - Seasons schedule caching
    - cache/index.ts - Clean export interface
  - Update all procedures to match new schema requirements
  - Skip 13th week series during regular weeks to prevent unnecessary API errors
  - Add comprehensive logging for debugging cache operations

  This implementation eliminates API errors from inactive special event series while
  maintaining complete data coverage during actual 13th week periods.
