--- drizzle/0001_real_wallflower.sql
@@ -1,77 +1,0 @@
-CREATE TABLE `race_schedule` (
-	`id` varchar(21) NOT NULL,
-	`season_id` int NOT NULL,
-	`race_week_num` int NOT NULL,
-	`schedule_name` varchar(150) NOT NULL,
-	`series_id` int NOT NULL,
-	`series_name` varchar(100) NOT NULL,
-	`category` varchar(25) NOT NULL,
-	`category_id` int NOT NULL,
-	`track_id` int NOT NULL,
-	`track_name` varchar(100) NOT NULL,
-	`special_event_type` int NOT NULL,
-	`session_minutes` int NOT NULL,
-	`temp_unit` int NOT NULL,
-	`temp_value` int NOT NULL,
-	`skies` int,
-	`wind_units` int NOT NULL,
-	`wind_value` int NOT NULL,
-	`time_of_day` int NOT NULL,
-	`track_water` int NOT NULL,
-	`allow_fog` boolean DEFAULT false,
-	`start_date` varchar(10) NOT NULL,
-	`first_sessoin_time` varchar(8) NOT NULL,
-	`repeat_minutes` int NOT NULL,
-	`week_end_time` varchar(30) NOT NULL,
-	`created_at` timestamp NOT NULL DEFAULT (now()),
-	`updated_at` timestamp NOT NULL DEFAULT (now()) ON UPDATE CURRENT_TIMESTAMP,
-	CONSTRAINT `race_schedule_id` PRIMARY KEY(`id`),
-	CONSTRAINT `uniqueSeasonWeek` UNIQUE(`season_id`,`race_week_num`)
-);
---> statement-breakpoint
-CREATE TABLE `season` (
-	`season_id` int NOT NULL,
-	`series_id` int NOT NULL,
-	`season_name` varchar(150) NOT NULL,
-	`season_short_name` varchar(50),
-	`season_year` int NOT NULL,
-	`season_quarter` int NOT NULL,
-	`active` boolean DEFAULT false,
-	`complete` boolean DEFAULT false,
-	`official` boolean DEFAULT true,
-	`license_group` int NOT NULL,
-	`cross_license` boolean DEFAULT false,
-	`max_weeks` int NOT NULL,
-	`race_week` int NOT NULL,
-	`drops` int NOT NULL,
-	`race_week_to_make_divisions` int NOT NULL,
-	`fixed_setup` boolean NOT NULL DEFAULT false,
-	`multiclass` boolean NOT NULL DEFAULT false,
-	`incident_limit` int NOT NULL,
-	`incident_warn_mode` int NOT NULL,
-	`incident_warn_param1` int NOT NULL,
-	`incident_warn_param2` int NOT NULL,
-	`connection_black_flag` boolean DEFAULT false,
-	`max_team_drivers` int NOT NULL,
-	`min_team_drivers` int NOT NULL,
-	`driver_changes` boolean DEFAULT false,
-	`driver_change_rule` int NOT NULL,
-	`qualifier_must_start_race` boolean DEFAULT false,
-	`grid_by_class` boolean DEFAULT false,
-	`op_duration` int NOT NULL,
-	`open_practice_session_type_id` int NOT NULL,
-	`num_fast_tows` int NOT NULL,
-	`num_opt_laps` int NOT NULL,
-	`schedule_description` varchar(200),
-	`start_date` varchar(30) NOT NULL,
-	`reg_user_count` int NOT NULL,
-	`created_at` timestamp NOT NULL DEFAULT (now()),
-	`updated_at` timestamp NOT NULL DEFAULT (now()) ON UPDATE CURRENT_TIMESTAMP,
-	CONSTRAINT `season_season_id` PRIMARY KEY(`season_id`),
-	CONSTRAINT `uniqueSeasonPeriod` UNIQUE(`series_id`,`season_year`,`season_quarter`)
-);
---> statement-breakpoint
-ALTER TABLE `user_chart_data` DROP INDEX `user_chart_data_user_id_category_chart_type_when_unique`;--> statement-breakpoint
-ALTER TABLE `user_chart_data` ADD CONSTRAINT `uniqueUserChartData` UNIQUE(`user_id`,`category`,`chart_type`,`when`);--> statement-breakpoint
-ALTER TABLE `race_schedule` ADD CONSTRAINT `race_schedule_season_id_season_season_id_fk` FOREIGN KEY (`season_id`) REFERENCES `season`(`season_id`) ON DELETE cascade ON UPDATE no action;--> statement-breakpoint
-ALTER TABLE `season` ADD CONSTRAINT `season_series_id_series_series_id_fk` FOREIGN KEY (`series_id`) REFERENCES `series`(`series_id`) ON DELETE no action ON UPDATE no action;
