--- drizzle/meta/_journal.json
@@ -1,20 +1,0 @@
-{
-  "version": "7",
-  "dialect": "mysql",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "5",
-      "when": 1756896344408,
-      "tag": "0000_melted_ultron",
-      "breakpoints": true
-    },
-    {
-      "idx": 1,
-      "version": "5",
-      "when": 1757411243736,
-      "tag": "0001_real_wallflower",
-      "breakpoints": true
-    }
-  ]
-}
