+++ src/app/api/cronjobs/cache/index.ts
@@ -1,0 +1,4 @@
+// Re-export all cache operations
+export { cacheSeries } from "./series";
+export { cacheWeeklyResults } from "./weekly-results";
+export { cacheSeasonsSchedule } from "./seasons-schedule";
