+++ src/app/api/cronjobs/route.ts
@@ -4,7 +4,7 @@ import { DateTime } from "luxon";
 import env from "@/env";
 
 import * as utilities from "@/app/api/cronjobs/utilities";
-import * as cacheOps from "@/app/api/cronjobs/cache-operations";
+import * as cacheOps from "@/app/api/cronjobs/cache";
 
 import { getOrRefreshAuthCode } from "@/modules/iracing/server/authentication";
 
@@ -21,28 +21,36 @@ export async function GET(request: NextRequest) {
     DateTime.now().toLocaleString(DateTime.DATETIME_MED),
   );
 
+  const seasonInfo = utilities.getCurrentSeasonInfo();
   const authCode = await getOrRefreshAuthCode();
 
-  const seriesCached = await cacheOps.cacheSeries({ authCode });
+  const seriesCacheParams = utilities.createSearchParams({
+    season_year: seasonInfo.currentYear,
+    season_quarter: seasonInfo.currentQuarter,
+  });
+  const seriesCached = await cacheOps.cacheSeries({
+    seasonYear: seasonInfo.currentYear,
+    seasonQuarter: seasonInfo.currentQuarter,
+    authCode,
+    seriesCacheParams,
+  });
 
   if (!seriesCached?.success) {
     console.error("Cron job error:", seriesCached?.error);
     return Response.json({ success: false }, { status: 500 });
   }
 
-  const seasonInfo = utilities.getCurrentSeasonInfo();
+  // const cachedSeasonsSchedule = await cacheOps.cacheSeasonsSchedule({
+  //   seasonYear: seasonInfo.currentYear,
+  //   seasonQuarter: seasonInfo.currentQuarter,
+  //   authCode,
+  // });
 
-  const cachedSeasonsSchedule = await cacheOps.cacheSeasonsSchedule({
-    seasonYear: seasonInfo.currentYear,
-    seasonQuarter: seasonInfo.currentQuarter,
-    authCode,
-  });
-
-  if (cachedSeasonsSchedule.success) {
-    console.log(`${cachedSeasonsSchedule.message}`);
-  } else {
-    console.warn(`${cachedSeasonsSchedule.message}`);
-  }
+  // if (cachedSeasonsSchedule.success) {
+  //   console.log(`${cachedSeasonsSchedule.message}`);
+  // } else {
+  //   console.warn(`YURR!!${cachedSeasonsSchedule.message}`);
+  // }
 
   const searchParams = utilities.createSearchParams({
     season_year: seasonInfo.currentYear,
@@ -57,6 +65,8 @@ export async function GET(request: NextRequest) {
     team_id: "",
     category_id: "",
   });
+
+  console.log("Calculated search params:", searchParams);
 
   const cachedWeeklyResults = await cacheOps.cacheWeeklyResults({
     authCode,
