+++ src/modules/iracing/server/procedures/get-all-series/index.ts
@@ -7,6 +7,7 @@ import { db } from "@/db";
 import { seriesTable, seriesWeeklyStatsTable } from "@/db/schemas";
 
 import { fetchData } from "@/modules/iracing/server/api";
+import { detectSpecialEventSeries } from "@/app/api/cronjobs/utilities";
 
 import { WeeklySeriesResultsInput } from "@/modules/iracing/server/procedures/weekly-series-results/schema";
 import { GetAllSeriesResponse } from "@/modules/iracing/server/procedures/get-all-series/schema";
@@ -72,11 +73,22 @@ export const cacheAllSeries = iracingProcedure.query(async ({ ctx }) => {
 
   const data = GetAllSeriesResponse.parse(res);
 
-  const insertValues = data.map((item) => ({
-    seriesId: item.series_id,
-    category: item.category,
-    seriesName: item.series_name,
-  }));
+  const insertValues = data.map((item) => {
+    const specialEventInfo = detectSpecialEventSeries(item.series_name);
+
+    return {
+      seriesId: item.series_id,
+      category: item.category,
+      categoryId: item.category_id,
+      seriesName: item.series_name,
+      seriesShortName: item.series_short_name,
+      eligible: item.eligible,
+      maxStarters: item.max_starters,
+      minStarters: item.min_starters,
+      isSpecialEvent: specialEventInfo.isSpecialEvent,
+      specialEventType: specialEventInfo.specialEventType,
+    };
+  });
 
   try {
     await db.transaction(async (tx) => {
